requirements.txt
altgraph==0.17.4
certifi==2021.5.30
charset-normalizer==2.0.12
distro==1.9.0
dnspython==2.2.1
idna==3.7
importlib-metadata==4.8.3
netifaces==0.11.0
pefile==2023.2.7
pip==21.2.2
psutil==6.0.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.0
pywin32==305
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.27.1
setuptools==58.0.4
typing_extensions==4.1.1
urllib3==1.26.19
wheel==0.37.1
wincertstore==0.2
zipp==3.6.0
xmltodict==0.13.0